Extract functions to their own file
