+++ worker/evaluate.js
@@ -1,0 +1,10 @@
+/**
+ * @param position {Int8Array}
+ */
+export const evaluate = (position) => {
+  let evaluation = 0;
+  for (var i = 0; i < position.length; i++) {
+    evaluation += position[i] ?? 0;
+  }
+  return evaluation;
+};
